pyproject.toml
[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sector-ai"
version = "1.1.0"
description = "Python Telegram Bot powered by LangChain and Ollama"
authors = [
    { name = "Rory Eckel" }
]
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.18",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "attrs==25.3.0",
    "certifi==2025.4.26",
    "chardet==5.2.0",
    "charset-normalizer==3.4.1",
    "cssselect2==0.8.0",
    "dataclasses-json==0.6.7",
    "filelock==3.18.0",
    "freetype-py==2.5.1",
    "frozenlist==1.6.0",
    "fsspec==2025.3.2",
    "greenlet==3.2.1",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "huggingface-hub==0.30.2",
    "idna==3.10",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==0.3.24",
    "langchain-core==0.3.56",
    "langchain-ollama==0.3.2",
    "langchain-text-splitters==0.3.8",
    "langsmith==0.3.37",
    "lxml==5.4.0",
    "marshmallow==3.26.1",
    "multidict==6.4.3",
    "mypy_extensions==1.1.0",
    "numpy==2.2.5",
    "ollama==0.4.8",
    "orjson==3.10.16",
    "packaging==24.2",
    "pillow==11.2.1",
    "propcache==0.3.1",
    "pycairo==1.28.0",
    "pydantic==2.11.3",
    "pydantic_core==2.33.1",
    "python-telegram-bot==22.0",
    "PyYAML==6.0.2",
    "regex==2024.11.6",
    "reportlab==4.4.0",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "rlPyCairo==0.3.0",
    "safetensors==0.5.3",
    "sniffio==1.3.1",
    "SQLAlchemy==2.0.40",
    "svglib==1.5.1",
    "tenacity==9.1.2",
    "tinycss2==1.4.0",
    "tokenizers==0.21.1",
    "tqdm==4.67.1",
    "transformers==4.51.3",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.0",
    "typing_extensions==4.13.2",
    "urllib3==2.4.0",
    "webencodings==0.5.1",
    "yarl==1.20.0",
    "zstandard==0.23.0",
]

[project.urls]
Homepage = "https://github.com/roryeckel/sector-ai"
Issues = "https://github.com/roryeckel/sector-ai/issues"

[tool.setuptools.packages.find]
where = ["src"]
